Rework some decoding functionality
Rather than adding panels with encoded text via JS, add them in the
template and simply decode them in JS. This prevents the email panel
pop-in on slow page loads while maintaining email address security.

Currently special cases are still made in the JS for email (to account
for the mailto: link text). Need to look into ways to more fully
templatize this and allow for more universally-applicable JS.