Install/configure Prezet, add sample post
The idea here is to just trash the overly complicated work I was doing
with a complex Laravel setup and just Keep It Simple, Stupid. Looked
into a couple options for markdown-based blogs and settled on this one
since it provides a lot of nice functionality out of the box while not
being too complicated or providing more functionality than I need. And
running a markdown-based blog makes things much easier to manage,
update, and port over if I decide to move to a different platform or
write my own.