"Better" linking of legacy resources
Builds were failing to properly pull in legacy vendor CSS/JS files. This
change generally fixes that, but was still failing to order the imports
of JS files properly and kept throwing jQuery undefined errors, so we're
just gonna use the CDN for now since this page as it is will hopefully
be relatively short-lived anyway.