"Better" filter routing
Use a URI and slug to filter to pass to the controller for filtering.
This currently limits filtering to either one category or tag at a time,
so it's a bit more limited than the query-string method, but it does
allow for better URI checking to highlight the nav menu.